rename java package and update bundle information
Package renamed from trace_event_logger to traceeventlogger

Update organization, name and url in pom.xml
