+++ src/main/java/org/eclipse/tracecompass/traceeventlogger/AsyncFileHandler.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 Ericsson
+ * Copyright (c) 2024, 2025 Ericsson
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the “Software”), to deal
@@ -22,7 +22,7 @@
  * SPDX-License-Identifier: MIT
  *******************************************************************************/
 
-package org.eclipse.tracecompass.trace_event_logger;
+package org.eclipse.tracecompass.traceeventlogger;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
@@ -41,7 +41,7 @@ import java.util.logging.LogManager;
 import java.util.logging.LogRecord;
 import java.util.logging.StreamHandler;
 
-import org.eclipse.tracecompass.trace_event_logger.LogUtils.TraceEventLogRecord;
+import org.eclipse.tracecompass.traceeventlogger.LogUtils.TraceEventLogRecord;
 
 /**
  * Asynchronous File Handler for JUL.
