+++ src/main/java/org/eclipse/tracecompass/traceeventlogger/InnerEvent.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 Ericsson
+ * Copyright (c) 2024, 2025 Ericsson
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the “Software”), to deal
@@ -21,11 +21,11 @@
  *
  * SPDX-License-Identifier: MIT
  *******************************************************************************/
-package org.eclipse.tracecompass.trace_event_logger;
+package org.eclipse.tracecompass.traceeventlogger;
 
 import java.util.logging.LogRecord;
 
-import org.eclipse.tracecompass.trace_event_logger.LogUtils.TraceEventLogRecord;
+import org.eclipse.tracecompass.traceeventlogger.LogUtils.TraceEventLogRecord;
 
 class InnerEvent {
