+++ src/main/java/org/eclipse/tracecompass/traceeventlogger/SnapshotHandler.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 Ericsson
+ * Copyright (c) 2024, 2025 Ericsson
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the “Software”), to deal
@@ -21,7 +21,7 @@
  *
  * SPDX-License-Identifier: MIT
  *******************************************************************************/
-package org.eclipse.tracecompass.trace_event_logger;
+package org.eclipse.tracecompass.traceeventlogger;
 
 import java.io.BufferedWriter;
 import java.io.File;
@@ -40,7 +40,7 @@ import java.util.logging.Level;
 import java.util.logging.LogManager;
 import java.util.logging.LogRecord;
 
-import org.eclipse.tracecompass.trace_event_logger.LogUtils.TraceEventLogRecord;
+import org.eclipse.tracecompass.traceeventlogger.LogUtils.TraceEventLogRecord;
 
 /**
  * Snapshot handler, will write to disk in a deferred way. Parameters to
