+++ src/test/java/org/eclipse/tracecompass/traceeventlogger/FailFilter.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 Ericsson
+ * Copyright (c) 2024, 2025 Ericsson
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the “Software”), to deal
@@ -22,7 +22,7 @@
  * SPDX-License-Identifier: MIT
  *******************************************************************************/
 
-package org.eclipse.tracecompass.trace_event_logger;
+package org.eclipse.tracecompass.traceeventlogger;
 
 import java.util.logging.Filter;
 import java.util.logging.LogRecord;
