+++ src/test/java/org/eclipse/tracecompass/traceeventlogger/SnapshotTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 Ericsson
+ * Copyright (c) 2024, 2025 Ericsson
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the “Software”), to deal
@@ -21,7 +21,7 @@
  *
  * SPDX-License-Identifier: MIT
  *******************************************************************************/
-package org.eclipse.tracecompass.trace_event_logger;
+package org.eclipse.tracecompass.traceeventlogger;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -197,7 +197,7 @@ public class SnapshotTest {
     @Test
     public void testConfigure() {
         try (InputStream fis = new FileInputStream(
-                new File("./src/test/java/org/eclipse/tracecompass/trace_event_logger/res/logging.properties"))) { //$NON-NLS-1$
+                new File("./src/test/java/org/eclipse/tracecompass/traceeventlogger/res/logging.properties"))) { //$NON-NLS-1$
             LogManager manager = LogManager.getLogManager();
             manager.readConfiguration(fis);
             Handler first = new SnapshotHandler();
@@ -215,7 +215,7 @@ public class SnapshotTest {
     @Test
     public void testGoodConfigure() {
         try (InputStream fis = new FileInputStream(
-                new File("./src/test/java/org/eclipse/tracecompass/trace_event_logger/res/goodlogging.properties"))) { //$NON-NLS-1$
+                new File("./src/test/java/org/eclipse/tracecompass/traceeventlogger/res/goodlogging.properties"))) { //$NON-NLS-1$
             LogManager manager = LogManager.getLogManager();
             manager.readConfiguration(fis);
             Handler first = new AsyncFileHandler(File.createTempFile("test", ".json").getAbsolutePath()); //$NON-NLS-1$ //$NON-NLS-2$
@@ -249,10 +249,10 @@ public class SnapshotTest {
     @Test
     public void testBadConfigure() {
         try (InputStream fis = new FileInputStream(
-                new File("./src/test/java/org/eclipse/tracecompass/trace_event_logger/res/badlogging.properties"))) { //$NON-NLS-1$
+                new File("./src/test/java/org/eclipse/tracecompass/traceeventlogger/res/badlogging.properties"))) { //$NON-NLS-1$
             LogManager manager = LogManager.getLogManager();
             manager.readConfiguration(fis);
-            String prop = manager.getProperty("org.eclipse.tracecompass.trace_event_logger.SnapshotHandler.maxEvents"); //$NON-NLS-1$
+            String prop = manager.getProperty("org.eclipse.tracecompass.traceeventlogger.SnapshotHandler.maxEvents"); //$NON-NLS-1$
             assertNotNull(prop);
             Handler first = new SnapshotHandler();
             first.close();
