+++ src/test/java/org/eclipse/tracecompass/traceeventlogger/res/badlogging.properties
@@ -1,6 +1,6 @@
 # MIT License
 # 
-# Copyright (c) 2024 Ericsson
+# Copyright (c) 2024, 2025 Ericsson
 # 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -23,15 +23,15 @@
 # SPDX-License-Identifier: MIT
 
 # used in AsyncFileHandlerTest as well as SnapshotTest
-handlers= org.eclipse.tracecompass.trace_event_logger.SnapshotHandler, org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler
+handlers= org.eclipse.tracecompass.traceeventlogger.SnapshotHandler, org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler
 
-org.eclipse.tracecompass.trace_event_logger.SnapshotHandler.maxEvents = -1
-org.eclipse.tracecompass.trace_event_logger.SnapshotHandler.timeout = -1
-org.eclipse.tracecompass.trace_event_logger.SnapshotHandler.filePath = 
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.maxSize = -1
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.queueDepth = -1
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.flushRate = -1
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.encoding = uTF-6
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.filter = coffeeFilter
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.errormanager = no thank you
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.formatter = LadyGaga
+org.eclipse.tracecompass.traceeventlogger.SnapshotHandler.maxEvents = -1
+org.eclipse.tracecompass.traceeventlogger.SnapshotHandler.timeout = -1
+org.eclipse.tracecompass.traceeventlogger.SnapshotHandler.filePath = 
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.maxSize = -1
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.queueDepth = -1
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.flushRate = -1
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.encoding = uTF-6
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.filter = coffeeFilter
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.errormanager = no thank you
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.formatter = LadyGaga
