+++ src/test/java/org/eclipse/tracecompass/traceeventlogger/res/benchmarklogging.properties
@@ -1,6 +1,6 @@
 # MIT License
 # 
-# Copyright (c) 2024 Ericsson
+# Copyright (c) 2024, 2025 Ericsson
 # 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -22,6 +22,8 @@
 #
 # SPDX-License-Identifier: MIT
 
-# used in AsyncFileHandlerTest as well as SnapshotTest
+# used in Benchmarks
 
-handlers= org.eclipse.tracecompass.trace_event_logger.SnapshotHandler, org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.maxSize = 1000
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.queueDepth = 100
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.flushRate = 100
