+++ src/test/java/org/eclipse/tracecompass/traceeventlogger/res/goodlogging.properties
@@ -1,6 +1,6 @@
 # MIT License
 # 
-# Copyright (c) 2024 Ericsson
+# Copyright (c) 2024, 2025 Ericsson
 # 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -24,15 +24,15 @@
 
 # used in AsyncFileHandlerTest as well as SnapshotTest
 
-handlers= org.eclipse.tracecompass.trace_event_logger.SnapshotHandler, org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler
+handlers= org.eclipse.tracecompass.traceeventlogger.SnapshotHandler, org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler
 
-org.eclipse.tracecompass.trace_event_logger.SnapshotHandler.maxEvents = 100
-org.eclipse.tracecompass.trace_event_logger.SnapshotHandler.timeout = 100
-org.eclipse.tracecompass.trace_event_logger.SnapshotHandler.filePath = "./traces/"
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.maxSize = 1000
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.queueDepth = 100
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.flushRate = 100
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.formatter = java.util.logging.SimpleFormatter
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.filter = org.eclipse.tracecompass.trace_event_logger.FailFilter
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.encoding = UTF-8
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.errormanager = java.util.logging.ErrorManager
+org.eclipse.tracecompass.traceeventlogger.SnapshotHandler.maxEvents = 100
+org.eclipse.tracecompass.traceeventlogger.SnapshotHandler.timeout = 100
+org.eclipse.tracecompass.traceeventlogger.SnapshotHandler.filePath = "./traces/"
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.maxSize = 1000
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.queueDepth = 100
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.flushRate = 100
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.formatter = java.util.logging.SimpleFormatter
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.filter = org.eclipse.tracecompass.traceeventlogger.FailFilter
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.encoding = UTF-8
+org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.errormanager = java.util.logging.ErrorManager
