+++ src/test/java/org/eclipse/tracecompass/traceeventlogger/res/logging.properties
@@ -1,6 +1,6 @@
 # MIT License
 # 
-# Copyright (c) 2024 Ericsson
+# Copyright (c) 2024, 2025 Ericsson
 # 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -22,8 +22,6 @@
 #
 # SPDX-License-Identifier: MIT
 
-# used in Benchmarks
+# used in AsyncFileHandlerTest as well as SnapshotTest
 
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.maxSize = 1000
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.queueDepth = 100
-org.eclipse.tracecompass.trace_event_logger.AsyncFileHandler.flushRate = 100
+handlers= org.eclipse.tracecompass.traceeventlogger.SnapshotHandler, org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler
